:art: refactor: Separate bot bootstrapping functions
